refactor(stats): 将统计数据类型从 int 改为 long
- 修改 PlugilyPlayerStatisticChangeEvent 中的 number 类型为 long- 更新 IStatsStorage 接口中的统计方法返回类型为 long
- 修改 IUser 接口中的统计相关方法参数和返回值为 long
- 更新 StatisticType 中 LEVEL、EXP 和 NEXT_LEVEL_EXP 的数据库字段类型为 bigint
- 修改 LeaderboardArgument 中的统计数据处理为 long 类型
- 更新 MessageBuilder 中的 integer 类型为 BigInteger 以支持大数值
- 修改 FileStats 和 MysqlManager 中的统计数据存储类型为 long- 更新 User 类中的统计数据映射类型为 Long
- 调整 UserManager 中的经验值计算以适应 long 类型

+++ MiniGamesBox API/src/main/java/plugily/projects/minigamesbox/api/events/player/PlugilyPlayerStatisticChangeEvent.java
@@ -35,9 +35,9 @@ public class PlugilyPlayerStatisticChangeEvent extends PlugilyEvent {
   private static final HandlerList HANDLERS = new HandlerList();
   private final Player player;
   private final IStatisticType statisticType;
-  private final int number;
+  private final long number;
 
-  public PlugilyPlayerStatisticChangeEvent(IPluginArena eventArena, Player player, IStatisticType statisticType, int number) {
+  public PlugilyPlayerStatisticChangeEvent(IPluginArena eventArena, Player player, IStatisticType statisticType, long number) {
     super(eventArena);
     this.player = player;
     this.statisticType = statisticType;
@@ -61,7 +61,7 @@ public class PlugilyPlayerStatisticChangeEvent extends PlugilyEvent {
     return statisticType;
   }
 
-  public int getNumber() {
+  public long getNumber() {
     return number;
   }
 }
